+++ src/eressea/editing.h
@@ -11,8 +11,8 @@
 #ifndef GM_EDITING
 #define GM_EDITING
 
-extern void make_block(short x, short y, short radius, const struct terrain_type * terrain);
-extern void make_island(short x, short y, int size);
+extern void make_block(int x, int y, int radius, const struct terrain_type * terrain);
+extern void make_island(int x, int y, int size);
 extern void seed_players(const char * filename, boolean new_island);
 
 #endif /* GM_EDITING */
